backend/requirements.txt
Flask
python-dotenv
firebase-admin
Flask-Cors